backend/requirements.txt
django
gunicorn
psycopg2-binary
djangorestframework
channels
channels_redis
daphne
Pillow